Level View
Changed SpawnLevelView to a more general LevelView to work with all levels.
Added the ability to change the level within LevelView.
Changed ViewManager to be a singleton.
Changed the way levels are loaded to no longer need to pass a path.
Created the ILevelChanger interface to create a way to callback level changing commands from level subclasses to the level view.