NPCs
Started work on npcs. Created the NPC class to generate an npc.
NPCS are randomly generated at the moment.
Added bounds and a way to look for entites by their bounds to the Entity class.
Created moveBounds variable in the mob class so npcs and the player can have specific movement bounds.
Renamed the levelRender method to renderUnderPlayer with the idea of also adding a renderOverPlayer method.
Added a solidEntities array to the level class that will hold all entities that are solid.
Collision method in Mob class now checks for both tiles and solid entities.