NPC Config
Created the NPCConfig class to pass config objects to npcs when creating them.
Created the BasicNPC class for creating a host of basic characters.
Added the renderTransparentSprite method to the screen class to allow sprites with alpha channels to be rendered.
Created a shadow under the guard player.
Updated the DialogBox class to auto close the dialog if the messages array is null.