Background Music Playback
Added playback methods to the Music class for playing, pausing, resuming and stopping the music.
Added a quit to main menu button to the pause screen.
Moved the music player reference from TestLevel class to the Level super class.
Stopped music playing when exiting level.
Added a destroyPlayer in GuardPlayer class to remove notifiers.
Added pauseLevel and resumeLevel methods to the LevelView class which is handing pausing music and playing pause menu sound effects.
Expanded the size of the pause menu to account for the new button.