Inventory System
Created the Item class to create static items with a public interface to get the specific values from them.
Created the InventoryManager class to perform CRUD operations on the inventory for a particular game instance.
Removed the escape key press notifier in preference of the general key press notifier.
Added the items sprite sheet.
Edited the Chest class to accept an Item in the constructor to act as the treasure of the chest.