Inventory Manager Operations
Removed the multiple methods for returning items of a certain type in preference of a single method that returns a filtered map of the inventory item map based on item type.
Wrote tests around the InventoryManager class.
Wrote tests around the Chest class.
Changed the background music for the test level.
Added the chest opening sound to chests.
Added two items of each type for testing.
Deleted a duplicate resource file.