Improved Lightmap Rendering
Changed the way the light map blends with the other sprites.
Lightmap is now applied on the fly with the rendering of sprites.
Lightmap entities are rendered before all sprites are drawn.
renderSprite method now takes another parameter to determine if it should be blended with the lightmap or not.
Changed to double buffering for faster rendering speed.