Chest Opening Animation
Added animation for when a chest is opened.
A chest can have an item or gold and the animation will play accordingly.
Changed the light map sprite for the lamp post.
Added a get and set gold method to the inventory manager class.
Wrote a test to make sure chests are increasing gold when opened.