Game Menu Nav Buttons
Added nav buttons to the game menu.
Added another font sheet to the font class.
Added a new method for rendering font characters allowing for characters to be rendered with different colors.
Made the button class more generic to allow for subclassing.
Created a PrimaryButton class to act as the menu buttons for the title screen and pause menu.
Created a GameMenuNavButton class for the nav buttons on the game menu.
Added a sprite for the background of the nav button to represent that it has been selected.