Begin Refactor To OpenGL
Added the JOGL library to the build.gradle dependencies.
Added a GLCanvas in place of the awt Canvas.
Created the EventListener class for receiving draw events for the GLCanvas.
Modified render methods to accept a GL2 parameter.
Refactored one of the renderSprite methods in the screen class to render a texture with OpenGL.
Added textures to sprites and sprite sheets.
Game is currently launching with no problems but there are more render methods that need to be refactored.