Walls
Added the wall class to create walls with a configuration object.
Added wall resources for basic walls.
Updated how collision detection works to work with skinny boundaries.
Added layering to improve how solid entities are displayed behind or in front of player.
Created the inside of the house on the test level with the new walls.