Level Scenes
Added support for levels to have different scenes where the music doesn't stop between them.
Added a method in the Music class to check if music is currently playing.
Added some comments to the Sprite class.
Added a second scene in the House Level.
Added a separate level for the upstairs before adding the scene support. This will be used for another area.
Created a Scene class to hold information about scenes.
Created the TransitionFader class to handle the fading between scenes and levels.
Added another constructor to the Portal class to accept portals that are intra level.
Adjusted the bounds on the wall so you can see the player through windows.