Upstairs Walls
Added walls to the upstairs section of the house.
Added the stairs to the upstairs section of the house.
Changed render under player and over player methods in Level class to under and over solid entities.
Fixed a bug that was causing the player to accept key input during transitions.