Game Menu Panels
Moved game menu tasks to different classes based on the panel they will be in charge of.
Added a new class for font styles.
Added a new constructor to font info class and added set color method.
Updated the look and feel of the nav buttons on the side panel.