Game Menu Views
Created the DisplayView class for creating instances of display views in the game menu.
Added a way for buttons on the game menu to switch views in the display panel.
Added placeholders for thumbnail images and added a reference to the file in the characters.json file.
Fixed an issue with the for each loop that was creating a warning in the console.