Party Member Buttons
Created the party member button class and added information about each character to the button.
Fixed a memory leak that was causing text rendering to slow down the application.
Added the GameColors class and the GameFonts class to act as a central location for font styles and colors.
Moved some files to new packages.