Code Comments
Added comments to several of the classes.
Added a condition in the updateStatBars method of the PartyMemberButton class to only set the new fill sprite if the value changed between updates.
Removed the if else condition in the enterPortal method of the Portal class since the check happens outside of the class.