Item Types
Created new classes for EquipmentItem, WeaponItem, ArmorItem and SpecialItem.
Changed the way the ItemBuilder class builds items so it returns the specific type.
Added a character field to the Item class with getters and setters.
Hard coded the InventoryListItem to set the first character in the party member array as the character for using items for testing.