Weapon Equipping
Added the ability to equip weapons to the player.
Weapons page on game menu now shows list of weapons.
Weapons page shows the current character and their stats.
Characters can now return stats based on what weapon they are equipped with.
Item detail card can now be closed out of with an x in the corner.