Swords and Weapon Sprite Sheet
Added the weapons sprite sheet with twelve types of each type of weapon.
Added the twelve swords to the game items.
Added a Sprite constructor to create scaled sprites from a sprite sheet.
Pushed the item attributes list down on the item detail card to allow more room for description.
Changed how equipment items get used.
Removed the item action attribute from items and removed that attribute from the item builder class.
Fixed a bug that was causing weapons to not show in the list if you replace your current weapon with another.