Pagination
Added pagination to the weapons view in the game menu.
Created the Pagination class for creating pagination on menu views.
Created the pagination button that is used by the Pagination class.
Added a constructor in the Button class to allow for buttons that have a consumer instead of a callback.
Exposed get top, right, bottom, and left in the Button class.
Added the ability to pass in font info to the game menu secondary button.
Changed the color of the equipped item slot background.
Added bow items to the game items.