Using Equipment Items
Added the ui flow for using healing items on characters.
The main view in the items view changes to a list of party members that you can select who to use the item on.
Renamed some variables in various files.
Created a new constructor in Sprite Class to create a scaled version of any existing sprite.