Use Stat Items
Added stats to the character details when using an equipment item.
When using a stat item the ui will update with the changes.
Prevented auto closing the select character screen on the items view until active animations are finished.
Added use sounds to equipment items.
Set the use sound for healing and mana restore items.
Turned off the hover sound on all buttons.