Save Files
Added files for saving the game.
Split up character information json into two parts. One part for base data and one for save data.
Created a Location class that is a spawn coordinate and level information.
Created a new game json file for default information when starting a game without a save.
Created a DataLoader class for loading save information.
Created a SaveDataManager for keeping track of save information while the game is being played.
Changed how the level keys in the level factory are getting set.
Made the level key and level name field a protected field in the Level class that gets set in the constructor of the sub classes.