Save Slot Buttons
Created save slot buttons so players can choose what slot they want to save their game to.
Added the saves folder to the git ignore.
Created images for the empty slot and empty slot when it is selected.
Make the initSaveObject method in the SaveDataManager public so when new games are started they can reset the save object to the new game settings.