New Player System
Created a new class to take the place of the Player class and its subclasses temporarily named PartyPlayer.
PartyPlayer class gets the current set party leader from the character manager and uses that information to display the appropriate character sheet for the player.
Fixed some sizing issues with the new thumbnail image.
Fixed a bug that was causing saving to break.