Character Battle Cards
Added battle cards to the bottom of the screen in the battle screen.
Cards show a thumbnail picture of the character, their health and mana bars.
Background of card slowly fills up with green as the characters ATB gauge fills up.
Slowed down the time it takes to fill up the ATB gauge for all levels.