Enemy Moves
Added a class for holding information about moves that enemies can make in battle.
Added the Bestiary class that will hold all of the monsters in the game as static resources.
Included an array of EnemyMove objects to the Enemy class and the EnemyBuilder class.
Added conditional checking on the moveFinished method to prevent bugs.
The moveFinished method now takes an argument to flag if the move was an offensive move to know if the enemy should end it's turn.