Enemy Move Builder
Created the EnemyMoveBuilder class for configuring EnemyMove objects.
Added move animations to the EnemyMove object and included it in the configuration.
Created the a test enemy in the bestiary with one move and a set animation.
EnemyBattleEntity now gets the next move from it's Enemy object and subsequently the animation for the move.