Battle Move Processor
Created the MoveProcessor class for processing battle moves.
Moved the EnemyMove and EnemyMoveBuilder classes to the new move package along with the MoveProcessor class.
Removed the handleAttack method from the CharacterBattleEntity class.
Fixed a bug that was throwing an error when the health or mana bars got to zero in the battle card. This could also be an issue in the game menu when showing the mana and health bars.