Flash Messages
Introduced flash messages that appear on the screen above the player when they are hit with an attack.
Randomized the attack power in the move processor so the same number isn't getting calculated for every attack.
Battle Entities now check their health and if they are at zero they enter a dead state.