Battle Move Controls
When a character is ready in battle the battle class will now call for the battle controls of that character to show.
Added the BattleControls class for displaying controls for the equipped moves of the given character.
Refactored the name of the EnemyMove and EnemyMoveBuilder classes to BattleMove and BattleMoveBuilder.
Added unique ids to all moves that are created in memory.
Added a couple more moves for testing.
Updated the processEnemyOffensiveMove method in the MoveProcessor class to handle magic attacks and physical attacks.
Added support for Characters to have an array of equipped moves.
Added an equipped move to each of the characters in the newGame.json for testing.
CharacterManager class now sets all the moves assigned to a character in the save file to the character object.