Character Experience
Added a method for getting the amount of experience needed to advance to the next level.
Added a method for increasing the experience of a character.
Fixed an issue that was causing stats to not increase when increaseLevel method was getting passed a value greater than one.
Created a test file for the Character class to test the methods around levels and experience.