Fixes #5 and Keylisteners
Fixed a bug that was crashing the battle when user clicks on an enemy before it is there turn.
Fixed an issue that was causing some key listeners to not get removed when moving between battling and not battling.
Added keyboard listeners to battle control buttons so the click sound does not play repeatedly on every update that the key is down.