Character Ring
Added an animated ring around the character who is currently selecting their move.
Changed the audio file for when a physical attack is missed.
Previously selected moves will be auto-selected the next time it is that characters turn.
Fixed a bug that was not setting the low health animation after enemies were getting attacked.