Battle Effects closes #3
Added more battle effects to the battle sequence.
Battle moves can be created with an animation for who it is targeting.
There is a delay on certain animations that get set for moves to make the timing a little more realistic.
Tried to fix a bug in the update travel method for the BattleEntity class that one time caused a null pointer exception.
Added two animations to be used with moves.