Heal Spell
Added a heal spell to the game for healing HP.
Added a new constructor to the flash message class to take in a target and it will randomise the x coordinate the flash message is shown in.
Added a setter in the flash message class for setting the color of the message.
Added a move information class for sending in a bundle of user, target, and move.
When defining defensive moves users can set the action in a method that happens when move is used.