Basic elemental types
Enemies and moves can now have elemental types.
When a magical attack move is used the attack power is modified based on the elemental type of enemy that it is targeting.
Made a small change to how the raw size of the battle sprite sheet is calculated for using different size sprite sheets for enemies.