Buff Moves
Added the ability to create buff moves for increasing stats in battle.
Created moves for defense buff and spirit buff and they are implemented in the move processor.
Added better constraints for the target selector for what battle entities can be selected based on the current move and the entity you are trying to select.