Multi-target moves: closes #26
Upgraded the battle system to allow for multi-targeting moves.
Characters and enemies can be equipped with multi-target moves.
Added a basic AI layer to enemies so they can more intelligently choose their next move and target.
When creating a monster the moves that are added can be given custom criteria for when they should be used and what weighted preference they have in relation to other moves.