Slot Data
Added slot data to the characters section of the newGame.json file.
When game is loaded the CharacterManager class loads in the slot information and sets up the equipped battle moves.
Added battle moves to the inventory of the newGame.json file.
The InventoryManager class now can add find battle moves to and from the inventory.