More Ability View Work
Added ability slots to the Character class.
Ability slots get created and added to the characters when the game loads.
Created an icon for displaying ability points and added ability points to the inventory.
Showing a lock icon on ability slots that have not been unlocked.
Showing a basic list of moves - just thumbnails - when ability slot is clicked on.
Made ability slot extend GroupButton so they could be added in a group.