Equip Item Moves
Users can now equip item moves from the abilities view in the game menu.
When the item move slot is upgraded and multiple moves can be equipped users can order which moves are equipped.
Added unequip functionality to all the item moves.
Hiding the upgrade button panel when the character is changed.