Party View
Started work on the party view in the game menu.
Added a component for showing the characters in the game with controls for adding and removing them from the party.
Added the party button to the button panel in the game menu.
Added some dummy data for three more characters to the characters.json, newGame.json and JSONUtil files.