Fixes #35
Fixed a bug that was causing the preview for saved games to show the party members out of order.
Made a change that now shows the characters in the game menu with their proper order.
Fixed an issue that was causing party members from previous game loads to show up in other games.