Crafting Package & Recipes
Added a new package called crafting to the top level for all classes related to crafting.
Added a recipe class that holds the individual recipes and a map of all of them.
Added the omelette item to the items file and updated the 32 x 32 sprite sheet to include a picture of an omelette.