Save / Load Refactor resolves #37
A transform layer is now in place to read and write to the file system then transform the data into JSON for the application to use.
All the logic for saving and loading has been divided between individual state managers who are in charge of their slice of the save data.
No class is aware of the file location or method of data loading / saving other than the transform layer.
Code is much more organized and reduced in many places.