Crafting Items Logic & Success Screen
Added the logic for crafting an output item from a recipe with the given input items.
Created a crafting item success screen that displays when an item is crafted.
Fixed some bugs on how the item view and itemSubView were displaying quantities.
Loaded recipes will now look at the crafted property to set if that recipe has ever been crafted.