QuestManager & Loading Quest Data
Added a QuestManager singleton for handling quest save data in the game.
Tested loading some quests from the newGame.json file.
Added some static quests in the Quests.java file for testing.
Fixed some typos in the Recipe.java file.
Fixed a bug where the tabs on the quest page in the game menu were not properly resetting.
Fixed a bug that was causing the enteringView method on game menu displayPanels to get called twice.