Working Quest Tabs & Pagination
Added more test quests to work on filtering different types of quests and showing them in the different tabs.
Added pagination for each of the tabs in the quest view on the game menu.
Quest list items now show an exclamation mark if they are new status.
Removed some unused system.out statements.